+++ src/features/widget/weather.ts
@@ -6,10 +6,6 @@ import { FeatureBase, type SettingLoader } from "../../libs/feature.js"
 import { Global } from "../../global.js"
 
 // #region Header
-namespace Header {   
-	export type Options = Partial<{
-	} & St.BoxLayout.ConstructorProps>
-}
 class Header extends St.BoxLayout {
 	_headerLabel: St.Label
 	_locationLabel: St.Label
@@ -40,6 +36,10 @@ class Header extends St.BoxLayout {
 	}
 }
 GObject.registerClass(Header)
+namespace Header {   
+	export type Options = Partial<{
+	} & St.BoxLayout.ConstructorProps>
+}
 // #endregion Header
 
 // #region WeatherWidget
@@ -55,6 +55,7 @@ class WeatherWidget extends St.BoxLayout {
 		super._init({
 			vertical: true,
 		})
+		this._updateStyleClass()
 
 		this.add_child(
 			this._header = new Header({})
@@ -89,14 +90,22 @@ class WeatherWidget extends St.BoxLayout {
 			GObject.BindingFlags.DEFAULT)
 		this._header._locationLabel.text = (this._item as any)._titleLocation.text
 	}
+	_updateStyleClass() {
+		const options = this._options
+		let style = "QSTWEAKS-weather"
+		if (options.removeShadow) style += " QSTWEAKS-weather-remove-shadow"
+		if (options.compact) style += " QSTWEAKS-weather-compact"
+		this.styleClass = style
+	}
 }
 GObject.registerClass(WeatherWidget)
 namespace WeatherWidget {
 	export interface Options {
 		clickCommand: string
+		compact: boolean
+		removeShadow: boolean
 	}
 }
-export { WeatherWidget }
 // #endregion WeatherWidget
 
 // #region WeatherWidgetFeature
@@ -117,19 +126,12 @@ export class WeatherWidgetFeature extends FeatureBase {
 	// #endregion settings
 
 	weatherWidget: WeatherWidget
-	updateStyleClass() {
-		let style = "QSTWEAKS-weather"
-		if (this.removeShadow) style += " QSTWEAKS-weather-remove-shadow"
-		if (this.compact) style += " QSTWEAKS-weather-compact"
-		this.weatherWidget.styleClass = style
-	}
-
 	override reload(key: string): void {
 		switch (key) {
 			case "weather-compact":
 			case "weather-remove-shadow":
 				if (!this.enabled) return
-				this.updateStyleClass()
+				this.weatherWidget!._updateStyleClass()
 				break
 			case "weather-click-command":
 				break
@@ -143,7 +145,6 @@ export class WeatherWidgetFeature extends FeatureBase {
 		this.maid.destroyJob(
 			this.weatherWidget = new WeatherWidget(this)
 		)
-		this.updateStyleClass()
 
 		Global.QuickSettingsGrid.add_child(this.weatherWidget)
 		Global.QuickSettingsGrid.layout_manager.child_set_property(